backend/requirements.txt
# backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
pydantic-settings==2.1.0
elasticsearch==8.12.0
requests==2.31.0
aiohttp==3.10.10